Migrates bindAws utility function into the common module since it's essentially a dupe
