+++ src/aws/ec2/vm/vm.js
@@ -12,7 +12,7 @@ const networkInterface = require('./../network-interface');
 const vmParams = require('./vm-params');
 const filter = require('./../ec2-filter');
 const util = require('../../../util');
-const awsUtil = require('../../aws-util');
+const awsCommon = require('../../common');
 
 module.exports = {
   bindAws,
@@ -46,7 +46,7 @@ module.exports = {
  * @returns {Object} this API bound to
  */
 function bindAws(aws) {
-  return awsUtil.bindAws(aws, module.exports);
+  return awsCommon.bindAws(aws, module.exports);
 }
 
 /**
